requirements.txt
aiogram
requests
flake8
flake8-import-order
flake8-variables-names
pep8-naming
flake8-builtins
flake8-bugbear
flake8-annotations
fastapi
uvicorn
psycopg2
sqlalchemy
alembic
pytest
pytest-asyncio
httpx
locust
fastapi-cache2[redis]
aioredis[standard]
celery
flower
aiohttp
xlwings
numpy
torch
Levenshtein